SL-1599: pull our logic for computing and preparing notes form listing-note-list component to notes-viewmodel-preparer service
