+++ src/app/hearing-part/components/listing-note-list/listing-note-list.component.ts
@@ -1,9 +1,9 @@
 import { Component, Input, OnInit, ViewChild } from '@angular/core';
 import { NoteListComponent } from '../../../notes/components/notes-list/note-list.component';
 import { Note } from '../../../notes/models/note.model';
-import { ListingCreateNotesConfiguration } from '../../models/listing-create-notes-configuration.model';
-import { getNoteViewModel, NoteViewmodel } from '../../../notes/models/note.viewmodel';
+import { NoteViewmodel } from '../../../notes/models/note.viewmodel';
 import { NotesPreparerService } from '../../../notes/services/notes-preparer.service';
+import { ListingNotesViewModelPreparerService } from '../../services/listing-notes-viewmodel-preparer.service';
 
 @Component({
   selector: 'app-listing-note-list',
@@ -18,26 +18,18 @@ export class ListingNoteListComponent implements OnInit {
   public noteViewModels: NoteViewmodel[] = [];
   public freeTextNoteViewModels: NoteViewmodel[] = [];
 
-  @Input() set notes(value: Note[]) {
-      this._notes = value;
-  }
-
-  @Input()
-  public entityId: string;
+  @Input() set notes(value: Note[]) { this._notes = value; }
+  @Input() public entityId: string;
 
   ngOnInit() {
       this.initiateNotes(this._notes);
   }
 
-  constructor(readonly listingNotesConfig: ListingCreateNotesConfiguration,
+  constructor(readonly viewModelPreparerService: ListingNotesViewModelPreparerService,
             public notePreparerService: NotesPreparerService) { }
 
   initiateNotes(notes: Note[]) {
-      this._notes = this.setNotesIfExist(notes);
-
-      this._notes.map(getNoteViewModel)
-          .map(this.disableShowingCreationDetailsOnNewNotes)
-          .map(this.makeExistingFreetextNotesReadonly)
+      this.viewModelPreparerService.prepare(notes)
           .forEach(this.disposeToProperArrays);
   }
 
@@ -45,58 +37,20 @@ export class ListingNoteListComponent implements OnInit {
       let preparedNotes = this.notePreparerService.prepare(
           this.noteList.getModifiedNotes(),
           this.entityId,
-          this.listingNotesConfig.entityName
+          this.viewModelPreparerService.notesConfig.entityName
       );
 
       let preparedFreeTextNotes = this.notePreparerService.prepare(
           this.freeTextNoteList.getModifiedNotes(),
           this.entityId,
-          this.listingNotesConfig.entityName
+          this.viewModelPreparerService.notesConfig.entityName
       );
       preparedFreeTextNotes = this.notePreparerService.removeEmptyNotes(preparedFreeTextNotes);
 
       return [...preparedNotes, ...preparedFreeTextNotes];
   }
 
-  protected setNotesIfExist(notes: Note[]) {
-      let defaultNotes = this.listingNotesConfig.defaultNotes().map(defaultNote => {
-          const alreadyExistingNote = notes.find(note => note.type === defaultNote.type);
-          if (alreadyExistingNote !== undefined) {
-              notes = notes.filter(n => n.id !== alreadyExistingNote.id);
-          }
-
-          return alreadyExistingNote || defaultNote
-      });
-
-      let combinedNotes = [...defaultNotes, ...notes];
-
-      if (this.containsOldFreeTextNote(combinedNotes)) {
-          return [this.listingNotesConfig.getNewFreeTextNote(), ...combinedNotes];
-      } else {
-          return combinedNotes;
-      }
-  }
-
-  protected containsOldFreeTextNote(notes: Note[]) {
-      return notes.find(n => n.type === 'Other note' && n.id === undefined) === undefined;
-  }
-
-  protected disableShowingCreationDetailsOnNewNotes(note: NoteViewmodel): NoteViewmodel {
-      if (note.id === undefined) {
-          note.displayCreationDetails = false;
-      }
-      return note;
-  }
-
-  protected makeExistingFreetextNotesReadonly(n: NoteViewmodel): NoteViewmodel {
-      if (n.type === 'Other note' && n.id !== undefined) {
-          n.readonly = true;
-      }
-
-      return n;
-  }
-
-  protected disposeToProperArrays = (n: NoteViewmodel) => {
+  private disposeToProperArrays = (n: NoteViewmodel) => {
       if (n.type === 'Other note') {
           this.freeTextNoteViewModels.push(n);
       } else {
